requirements.txt
aerich==0.7.2
aiodns==3.0.0
aiohttp==3.8.5
aiosignal==1.3.1
aiosqlite==0.17.0
annotated-types==0.5.0
anyio==3.7.1
appdirs==1.4.4
async-timeout==4.0.3
asyncio==3.4.3
asyncpg==0.28.0
attrs==23.1.0
beautifulsoup4==4.12.2
Brotli==1.1.0
bs4==0.0.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
dictdiffer==0.9.0
exceptiongroup==1.1.3
fastapi==0.103.1
frozenlist==1.4.0
greenlet==2.0.2
h11==0.14.0
httptools==0.6.0
idna==3.4
iso8601==1.1.0
jedi==0.19.0
lxml==4.9.3
multidict==6.0.4
parso==0.8.3
playwright==1.37.0
prompt-toolkit==3.0.39
psycopg==3.1.10
psycopg-binary==3.1.10
psycopg-pool==3.1.7
psycopg2-binary==2.9.7
ptpython==3.0.23
pycares==4.3.0
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
pyee==9.0.4
Pygments==2.16.1
pypika-tortoise==0.1.6
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
sniffio==1.3.0
soupsieve==2.5
starlette==0.27.0
termcolor==2.3.0
tomlkit==0.12.1
tortoise-cli==0.1.2
tortoise-orm==0.20.0
typing_extensions==4.7.1
ua-generator==0.1.8
urllib3==2.0.4
uvicorn==0.23.2
uvloop==0.17.0
validators==0.22.0
watchfiles==0.20.0
wcwidth==0.2.6
websockets==11.0.3
yarl==1.9.2
